+++ lifecycle/lifecycle/django/registry/models.py
@@ -82,6 +82,7 @@ class Deployment(models.Model):
     phase = models.TextField(null=True, blank=True)
     image_name = models.TextField(null=True, blank=True)
     infrastructure_target = models.CharField(max_length=256, null=True)
+    warnings = models.TextField(null=True, blank=True)
 
     def __str__(self):
         return f'{self.id}'