+++ racetrack_client/racetrack_client/client/deploy.py
@@ -113,10 +113,9 @@ def send_deploy_request(
         logger.info(f'job deployment requested: {deploy_id}')
 
         try:
-            job_url = _wait_for_deployment_result(lifecycle_url, deploy_id, user_auth, [])
+            job_url = _wait_for_deployment_result(lifecycle_url, deploy_id, user_auth, [], [])
         except Exception as e:
             raise DeploymentError(e)
-
         logger.info(f'Job "{manifest.name}" has been deployed. Check out {job_url} to access your Job')
     finally:
         if tmp_file is not None and os.path.exists(tmp_file) and os.path.isfile(tmp_file):
@@ -144,13 +143,19 @@ def get_deploy_request_payload(
 @backoff.on_exception(
     backoff.fibo, TimeoutError, max_value=3, max_time=DEPLOYMENT_TIMEOUT_SECS, jitter=None, logger=None
 )
-def _wait_for_deployment_result(lifecycle_url: str, deploy_id: str, user_auth: str, phases: List[Optional[str]]) -> str:
+def _wait_for_deployment_result(lifecycle_url: str, deploy_id: str, user_auth: str, phases: List[Optional[str]], warningss: List[Optional[str]]) -> str:
     # see `lifecycle.endpoints.deploy::setup_deploy_endpoints::DeployIdEndpoint` for server-side implementation
     r = Requests.get(
         f'{lifecycle_url}/api/v1/deploy/{deploy_id}',
         headers=get_auth_request_headers(user_auth),
     )
     response = parse_response_object(r, 'Lifecycle deployment status')
+    warnings = response['warnings']
+    if not warningss or warningss[-1] != warnings:  # don't print the same warnings again
+        warningss.append(warnings)
+        if warnings:
+            logger.warning(warnings)
+
     status = response['status'].lower()
     if status == 'failed':
         raise RuntimeError(response['error'])
@@ -164,6 +169,8 @@ def _wait_for_deployment_result(lifecycle_url: str, deploy_id: str, user_auth: s
             logger.info(f'deployment in progress: {phase}...')
         else:
             logger.info(f'deployment in progress...')
+
+
     raise TimeoutError('Deployment timeout error')